requirements.txt
attrs==21.4.0
banal==1.0.6
beautifulsoup4==4.11.1
binaryornot==0.4.4
boolean.py==3.8
certifi==2024.7.4
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
commoncode==30.2.0
construct==2.10.68
container-inspector==31.0.0
cryptography==43.0.1
python-dateutil==2.8.2
debian-inspector==30.0.0
dockerfile-parse==1.2.0
dparse2==0.6.1
extractcode==31.0.0
extractcode-7z==16.5.210531
extractcode-libarchive==3.5.1.210531
fasteners==0.17.3
fingerprints==1.0.3
ftfy==6.0.3
future==0.18.3
gemfileparser==0.8.0
html5lib==1.1
idna==3.7
importlib-metadata==4.8.3
inflection==0.5.1
intbitset==3.1.0
isodate==0.6.1
jaraco.functools==3.4.0
javaproperties==0.8.1
Jinja2==3.1.4
jsonstreams==0.6.0
license-expression==21.6.14
lxml==4.9.1
MarkupSafe==2.0.1
more-itertools==8.13.0
normality==2.3.3
packagedcode-msitools==0.101.210706
packageurl-python==0.9.9
packaging==21.3
parameter-expansion-patched==0.3.1
patch==1.16
pdfminer-six==20220506
pefile==2021.9.3
pip-requirements-parser==31.2.0
pkginfo2==30.0.0
pluggy==1.0.0
plugincode==30.0.0
ply==3.11
publicsuffix2==2.20191221
pyahocorasick==2.0.0b1
pycparser==2.21
pygmars==0.7.0
Pygments==2.15.0
pymaven-patch==0.3.0
pyparsing==3.0.8
pytz==2022.1
PyYAML==6.0
rdflib==5.0.0
regipy==2.3.1
requests==2.32.0
rpm-inspector-rpm==4.16.1.3.210404
saneyaml==0.5.2
six==1.16.0
soupsieve==2.3.1
spdx-tools==0.7.0a3
text-unidecode==1.3
toml==0.10.2
typecode==30.0.0
typecode-libmagic==5.39.210531
univers==30.11.0
urllib3==1.26.19
urlpy==0.5
wcwidth==0.2.5
webencodings==0.5.1
xmltodict==0.12.0
zipp==3.19.1
